+++ src/test/java/eu/bbmri/eric/csit/service/negotiator/unit/service/AccessCriteriaSetServiceTest.java
@@ -4,7 +4,6 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
-import eu.bbmri.eric.csit.service.negotiator.NegotiatorApplication;
 import eu.bbmri.eric.csit.service.negotiator.database.repository.AccessCriteriaSetRepository;
 import eu.bbmri.eric.csit.service.negotiator.exceptions.EntityNotFoundException;
 import eu.bbmri.eric.csit.service.negotiator.service.AccessCriteriaSetServiceImpl;
@@ -12,26 +11,26 @@ import java.util.Optional;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.ActiveProfiles;
+import org.modelmapper.ModelMapper;
 
-@SpringBootTest(classes = NegotiatorApplication.class)
-@ActiveProfiles("test")
 public class AccessCriteriaSetServiceTest {
 
   @Mock
   AccessCriteriaSetRepository accessCriteriaSetRepository;
 
-  @Autowired
-  AccessCriteriaSetServiceImpl accessCriteriaSetService;
+  @Mock
+  ModelMapper modelMapper;
+
+  @InjectMocks
+  AccessCriteriaSetServiceImpl service;
 
   private AutoCloseable closeable;
 
   @BeforeEach
-  void beforeAll() {
+  void before() {
     closeable = MockitoAnnotations.openMocks(this);
   }
 
@@ -44,7 +43,7 @@ public class AccessCriteriaSetServiceTest {
   void testRaiseException_whenAccessCriteriaNotFound() {
     when(accessCriteriaSetRepository.findByResourceEntityId(any())).thenReturn(Optional.empty());
     assertThrows(EntityNotFoundException.class,
-        () -> accessCriteriaSetService.findByResourceEntityId("aResourceId"));
+        () -> service.findByResourceEntityId("aResourceId"));
 
   }
 }
